+++ src/main/java/mathano/mathano/handlers/KitsGui.java
@@ -1,6 +1,5 @@
 package mathano.mathano.handlers;
 
-import com.google.gson.Gson;
 import dev.triumphteam.gui.builder.item.ItemBuilder;
 import dev.triumphteam.gui.guis.Gui;
 import dev.triumphteam.gui.guis.GuiItem;
@@ -8,12 +7,8 @@ import dev.triumphteam.gui.guis.PaginatedGui;
 import dev.triumphteam.gui.guis.StorageGui;
 import mathano.mathano.OBGiveAll;
 import mathano.mathano.database.DataKits;
-import mathano.mathano.database.jsondata.DataKitsJson;
-import mathano.mathano.database.serialization.Serialization;
 import mathano.mathano.enums.Placeholders;
 import mathano.mathano.managers.DataKitsManager;
-import mathano.mathano.managers.DatabaseManager;
-import mathano.mathano.managers.JsonManager;
 import mathano.mathano.managers.RewardsManager;
 import mathano.mathano.utils.ItemGui;
 import mathano.mathano.utils.Utils;
@@ -21,13 +16,11 @@ import net.kyori.adventure.text.Component;
 import net.wesjd.anvilgui.AnvilGUI;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
-import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 
-import java.io.IOException;
 import java.util.*;
 
 public class KitsGui {
@@ -77,7 +70,7 @@ public class KitsGui {
         mainGui.setItem(6, 8, glassPaneItemGui);
         mainGui.setItem(6, 9, exitGuiItem);
 
-        DatabaseManager.dataKits.forEach((kit_name, dataKit) -> {
+        DataKitsManager.dataKits.forEach((kit_name, dataKit) -> {
             GuiItem iconGui = ItemBuilder.from(dataKit.getIcon()).asGuiItem(inventoryClickEvent -> {
                 kitEditGUI(player, inventoryClickEvent.getCurrentItem().getItemMeta().getDisplayName());
             });
@@ -111,7 +104,6 @@ public class KitsGui {
             }
 
             if (length > 0) {
-                int finalLength = length;
                 new AnvilGUI.Builder()
                         // Either use sync or async variant, not both
                         .onClick((slot, stateSnapshot) -> {
@@ -119,7 +111,7 @@ public class KitsGui {
                                 return Collections.emptyList();
                             }
 
-                            return verifKitName(stateSnapshot.getText(), null, inventoryClickEvent.getClickedInventory(), player, false, finalLength);
+                            return verifKitName(stateSnapshot.getText(), null, inventoryClickEvent.getClickedInventory(), player, false);
                         })
                         // Sets the text the GUI should start with
                         .text("Nom du kit")
@@ -179,7 +171,6 @@ public class KitsGui {
             }
 
             if (length > 0) {
-                int finalLength = length;
                 new AnvilGUI.Builder()
                         // Either use sync or async variant, not both
                         .onClick((slot, stateSnapshot) -> {
@@ -187,7 +178,7 @@ public class KitsGui {
                                 return Collections.emptyList();
                             }
 
-                            return verifKitName(stateSnapshot.getText(), name, inventoryClickEvent.getClickedInventory(), player, true, finalLength);
+                            return verifKitName(stateSnapshot.getText(), name, inventoryClickEvent.getClickedInventory(), player, true);
                         })
                         // Sets the text the GUI should start with
                         .text(name)
@@ -216,19 +207,7 @@ public class KitsGui {
 
             GuiItem confirmItemGui = ItemBuilder.from(ItemGui.confirmItem).asGuiItem(inventoryClickEvent1 -> {
 
-
-
-                // VERIFIER SI FONCTIONNE AVEC REMPLACEMENT PAR name
-                //
-                //
-                // IMPORTANT
-                deleteKit(inventoryClickEvent.getClickedInventory().getItem(47).getItemMeta().getDisplayName(), player);
-
-
-
-
-
-
+                deleteKit(name, player);
 
                 mainGui(player);
             });
@@ -257,7 +236,7 @@ public class KitsGui {
         kitEditGui.setItem(6, 8, glassPaneItemGui);
         kitEditGui.setItem(6, 9, exitGuiItem);
 
-        DataKits dataKit = DatabaseManager.dataKits.get(name);
+        DataKits dataKit = DataKitsManager.dataKits.get(name);
 
         kitEditGui.addItem(dataKit.getItems());
 
@@ -266,138 +245,46 @@ public class KitsGui {
         kitEditGui.open(player);
     }
 
-    // Save the kit in the cache, handles renaming
-    public static void saveKit(Inventory kit, Player player, String name, String oldName, int length) {
-        // init variables
-        ItemStack[] items = new ItemStack[length];
-        ItemStack icon;
-        ItemMeta iconMeta;
-        int cmp = 0;
-        boolean edit = false;
-
-        if (kit.getItem(47) != null) {
-            icon = kit.getItem(47);
-        } else {
-            icon = new ItemStack(Material.CHEST);
-        }
-
-        iconMeta = icon.getItemMeta();
-        iconMeta.setDisplayName(name);
-        icon.setItemMeta(iconMeta);
-
-        if (DataKitsManager.DATA_KITS_CONFIG.contains(name)) {
-            DataKitsManager.DATA_KITS_CONFIG.set(name, null);
-            edit = true;
-        }
-
-        if (oldName != null && DataKitsManager.DATA_KITS_CONFIG.contains(oldName)) {
-            DataKitsManager.DATA_KITS_CONFIG.set(oldName, null);
-
-            for (String key : RewardsManager.REWARDS_CONFIG.getKeys(false)) {
-                ConfigurationSection section = RewardsManager.REWARDS_CONFIG.getConfigurationSection(key);
-
-                if (section.contains(oldName)) {
-                    int numberKits = section.getInt(oldName);
-                    section.set(name, numberKits);
-
-                    section.set(oldName, null);
-                    if (section.getKeys(false).size() <= 0) {
-                        RewardsManager.REWARDS_CONFIG.set(player.getUniqueId().toString(), null);
-                    }
-                }
-            }
-
-            edit = true;
-        }
-
-        DataKitsManager.DATA_KITS_CONFIG.set(name + ".name", icon);
-
-        for (int i = 0; i < 36; i++) {
-            if (kit.getItem(i) != null) {
-                items[cmp] = kit.getItem(i);
-                cmp++;
-            }
-        }
-
-        for (int i = 0; i < cmp; i++) {
-            DataKitsManager.DATA_KITS_CONFIG.set(name + "." + i, items[i]);
-        }
-
-        if (edit) {
-            // Message sent whenever a kit was successfully modified
-            player.sendMessage(Utils.getText(section, "kitModified", Placeholders.KIT_NAME.set(name)));
-
-        } else {
-            // Message sent whenever a kit was successfully created
-            player.sendMessage(Utils.getText(section, "kitCreated", Placeholders.KIT_NAME.set(name)));
-        }
-
-        DataKitsManager.INSTANCE.save();
-    }
-
     // Deletes the selected kit from the DataKits and Rewards cache
     public static void deleteKit(String name, Player player) {
-        DatabaseManager.dataKits.remove(name);
+        DataKitsManager.dataKits.remove(name);
         player.sendMessage(ChatColor.GREEN + "Le kit " + name + " a été supprimé !");
 
-
-
-
-
-
-
-
-        // for loop pour adapter rewards !!!!!
-
-
-
-
-
-
-
-
-        /*DataKitsManager.DATA_KITS_CONFIG.set(name, null);
-        player.sendMessage(ChatColor.GREEN + "Le kit " + name + " a été supprimé !");
-
-        for (String key : RewardsManager.REWARDS_CONFIG.getKeys(false)) {
-            ConfigurationSection section = RewardsManager.REWARDS_CONFIG.getConfigurationSection(key);
-
-            if (section.contains(name)) {
-                section.set(name, null);
-                if (section.getKeys(false).size() <= 0) {
-                    RewardsManager.REWARDS_CONFIG.set(player.getUniqueId().toString(), null);
+        RewardsManager.rewards.forEach((uuid, rewardsHash) -> {
+            if (rewardsHash.containsKey(name)) {
+                rewardsHash.remove(name);
+                if (rewardsHash.size() <= 0) {
+                    RewardsManager.rewards.remove(uuid);
+                } else {
+                    RewardsManager.rewards.put(uuid, rewardsHash);
                 }
             }
-        }*/
+        });
     }
 
-    public List<AnvilGUI.ResponseAction> verifKitName(String newKitName, String oldKitName, Inventory clickedInventory, Player player, Boolean edit, int length) {
+    public List<AnvilGUI.ResponseAction> verifKitName(String newKitName, String oldKitName, Inventory clickedInventory, Player player, Boolean edit) {
         if (!newKitName.equalsIgnoreCase("")
                 && !newKitName.contains(" ")) {
             if (!edit) {
 
                 //DataKitsManager.DATA_KITS_CONFIG.contains(newKitName)
-                if (DatabaseManager.dataKits.containsKey(newKitName)) {
+                if (DataKitsManager.dataKits.containsKey(newKitName)) {
                     // Message sent when the saved kit already exists
                     player.sendMessage(Utils.getText(section, "alreadyExists", Placeholders.KIT_NAME.set(newKitName)));
                     return Arrays.asList(AnvilGUI.ResponseAction.replaceInputText("Nom du kit"));
                 } else {
-                    //saveKit(clickedInventory, player, newKitName, null, length);
-                    saveKit2(clickedInventory, player, newKitName, null);
+                    saveKit(clickedInventory, player, newKitName, null);
                     return Arrays.asList(AnvilGUI.ResponseAction.close());
                 }
             } else {
-                // DataKitsManager.DATA_KITS_CONFIG.contains(newKitName)
-                if (DatabaseManager.dataKits.containsKey(newKitName) && newKitName.equals(oldKitName)) {
-                    //saveKit(clickedInventory, player, newKitName, DataKitsManager.DATA_KITS_CONFIG.getConfigurationSection(oldKitName).getItemStack("name").getItemMeta().getDisplayName(), length);
-                    saveKit2(clickedInventory, player, newKitName, oldKitName);
+                if (DataKitsManager.dataKits.containsKey(newKitName) && newKitName.equals(oldKitName)) {
+                    saveKit(clickedInventory, player, newKitName, oldKitName);
                     return Arrays.asList(AnvilGUI.ResponseAction.close());
                 }
 
                 // !DataKitsManager.DATA_KITS_CONFIG.contains(newKitName)
-                if (!DatabaseManager.dataKits.containsKey(newKitName)) {
-                    //saveKit(clickedInventory, player, newKitName, DataKitsManager.DATA_KITS_CONFIG.getConfigurationSection(oldKitName).getItemStack("name").getItemMeta().getDisplayName(), length);
-                    saveKit2(clickedInventory, player, newKitName, oldKitName);
+                if (!DataKitsManager.dataKits.containsKey(newKitName)) {
+                    saveKit(clickedInventory, player, newKitName, oldKitName);
                     return Arrays.asList(AnvilGUI.ResponseAction.close());
                 } else {
                     player.sendMessage(ChatColor.RED + "Le kit " + newKitName + " existe déjà !");
@@ -413,20 +300,8 @@ public class KitsGui {
         }
     }
 
-    private List<ItemStack> getKitContent(Inventory kit) {
-        List<ItemStack> items = new ArrayList<>();
-        ItemStack item;
-
-        for (int i = 0; i < 36; i++) {
-            item = kit.getItem(i);
-            if(item == null)continue;
-
-            items.add(item);
-        }
-        return items;
-    }
-
-    private void saveKit2(Inventory kit, Player player, String name, String oldName) {
+    // Save the kit in the cache, handles renaming
+    private void saveKit(Inventory kit, Player player, String name, String oldName) {
         boolean edit = false;
 
         ItemStack icon = kit.getItem(47);
@@ -438,7 +313,22 @@ public class KitsGui {
         iconMeta.setDisplayName(name);
         icon.setItemMeta(iconMeta);
 
-        if (DatabaseManager.dataKits.containsKey(name)) {
+        if (DataKitsManager.dataKits.containsKey(name)) {
+            DataKitsManager.dataKits.remove(name);
+            edit = true;
+        }
+
+        if (oldName != null && DataKitsManager.dataKits.containsKey(oldName)) {
+            DataKitsManager.dataKits.remove(oldName);
+
+            RewardsManager.rewards.forEach((uuid, rewardsHash) -> {
+                if (rewardsHash.containsKey(oldName)) {
+                    rewardsHash.put(name, rewardsHash.get(oldName));
+                    rewardsHash.remove(oldName);
+                    RewardsManager.rewards.put(uuid, rewardsHash);
+                }
+            });
+
             edit = true;
         }
 
@@ -447,26 +337,7 @@ public class KitsGui {
         dataKit.setName(name);
         dataKit.setIcon(icon);
         dataKit.setItems(getKitContent(kit));
-        DatabaseManager.dataKits.put(name, dataKit);
-
-
-        if (oldName != null && DatabaseManager.dataKits.containsKey(oldName)) {
-            DatabaseManager.dataKits.remove(oldName);
-
-
-
-
-
-            // for loop pour adapter les rewards !!!!!
-
-
-
-
-
-
-            edit = true;
-        }
-
+        DataKitsManager.dataKits.put(name, dataKit);
 
         if (edit) {
             // Message sent whenever a kit was successfully modified
@@ -476,68 +347,18 @@ public class KitsGui {
             // Message sent whenever a kit was successfully created
             player.sendMessage(Utils.getText(section, "kitCreated", Placeholders.KIT_NAME.set(name)));
         }
+    }
 
+    private List<ItemStack> getKitContent(Inventory kit) {
+        List<ItemStack> items = new ArrayList<>();
+        ItemStack item;
 
+        for (int i = 0; i < 36; i++) {
+            item = kit.getItem(i);
+            if(item == null)continue;
 
-
-
-        /*Gson gson = new Gson();
-        String iconJson;
-
-        if (icon == null) {
-            icon = new ItemStack(Material.CHEST);
+            items.add(item);
         }
-        ItemMeta iconMeta = icon.getItemMeta();
-        iconMeta.setDisplayName(name);
-
-        icon.setItemMeta(iconMeta);
-
-        //iconJson.setItemStack(icon);
-        iconJson = Serialization.INSTANCE.serializeAndEncodeItemStack(icon);
-
-        //List<ItemStackJson> itemStackJsonList = new ArrayList<>();
-        List<String> itemStackJsonList = new ArrayList<>();
-
-        getKitContent(kit).forEach(itemStack -> {
-            //ItemStackJson item = new ItemStackJson();
-            //item.setItemStack(itemStack);
-            //itemStackJsonList.add(item);
-            itemStackJsonList.add(Serialization.INSTANCE.serializeAndEncodeItemStack(itemStack));
-        });
-
-        DataKitsJson dataKitsJson = new DataKitsJson();
-        dataKitsJson.setIcon(iconJson);
-        dataKitsJson.setItems(itemStackJsonList);
-        dataKitsJson.setName(name);
-
-        OBGiveAll.INSTANCE.getLogger().info(JsonManager.INSTANCE.createJsonKit(dataKitsJson));
-
-        OBGiveAll.INSTANCE.getLogger().info("");
-
-        //DatabaseManager.INSTANCE.addKit(dataKitsJson);
-
-        if (DatabaseManager.dataKits.get(name) != null) {
-
-        }
-
-        try {
-            OBGiveAll.INSTANCE.getLogger().info(Serialization.INSTANCE.deserialize(JsonManager.INSTANCE.reader.readValue(JsonManager.INSTANCE.createJsonKit(dataKitsJson), DataKitsJson.class).getIcon()));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        try {
-            String content = DatabaseManager.INSTANCE.getKit(dataKitsJson.getName());
-
-            List<ItemStack> itemStackList = Serialization.INSTANCE.deserializeAndDecodeItemStackList(JsonManager.INSTANCE.reader.readValue(content, DataKitsJson.class).getItems());
-
-            itemStackList.forEach(itemStack -> {
-                player.getInventory().addItem(itemStack);
-                player.sendMessage(Objects.requireNonNull(itemStack.getItemMeta()).getDisplayName() + " obtenu avec bdd");
-            });
-
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }*/
+        return items;
     }
 }
