+++ src/main/resources/config.yml
@@ -1,3 +1,4 @@
+# url: <ip|domain>:port/databaseName
 database:
   url: ""
   username: ""
